Added dGPU cold state check
TL;DR : Reduces unwanted power usage (especially when on battery)
Since checking the dGPU temperature wakes it up from its cold state, it causes unwanted power usage when the dGPU is not in use. So now it first checks whether its in D3Cold state.
Also there is a added 8 second cool down that is triggered when GPU usage is 0 so it can reenter the cold state.